Move more form logic into the base changeset
